Separate and refactor type matching machinery
